スコアの送信にtoken認証を必要とするようにした
