+++ build.gradle
@@ -22,7 +22,7 @@ dependencies {
     compileOnly 'org.projectlombok:lombok:1.16.16'
     // This dependency is used internally, and not exposed to consumers on their own compile classpath.
     implementation 'com.google.guava:guava:23.0'
-    // compile('commons-codec:commons-codec:1.11')
+    compile('commons-codec:commons-codec:1.11')
     // compile 'com.squareup.retrofit2:retrofit:2.1.0'
     // compile 'com.squareup.retrofit2:converter-gson:2.1.0'
     // compile 'com.squareup.okhttp3:logging-interceptor:3.3.1'
