+++ src/main/java/config/ScoreConfig.java
@@ -3,4 +3,5 @@ package config;
 public class ScoreConfig {
     public static final int enemyBreakPoint = 100;
     public static final String serverApiUrl = "http://localhost:8080/api/score/";
+    public static final int md5 = 11;
 }
